New way of doing the iteration
With the method ord we get the integer ordinal of a one character string,
in other word we get the ascii decimal code of that character.
With that we can do the encryptation by adding the code with the shift.
The only problem that i have right now is adding the login when the shift goes
beyond the letter z